+++ extension/chrome/settings/modules/add_key_generate_module.ts
@@ -13,7 +13,8 @@ import { OpenPGPKey } from '../../../js/common/core/crypto/pgp/openpgp-key.js';
 import { saveKeysAndPassPhrase } from '../../../js/common/helpers.js';
 import { submitPublicKeyIfNeeded } from '../../../js/common/key-helper.js';
 import { Lang } from '../../../js/common/lang.js';
-import { Catch, CompanyLdapKeyMismatchError } from '../../../js/common/platform/catch.js';
+import { Catch } from '../../../js/common/platform/catch.js';
+import { CompanyLdapKeyMismatchError } from '../../../js/common/platform/error-report.js';
 import { AcctStore } from '../../../js/common/platform/store/acct-store.js';
 import { KeyStore } from '../../../js/common/platform/store/key-store.js';
 import { Xss } from '../../../js/common/platform/xss.js';
