+++ extension/chrome/settings/setup.ts
@@ -16,7 +16,8 @@ import { KeyStoreUtil } from '../../js/common/core/crypto/key-store-util.js';
 import { KeyInfoWithIdentity } from '../../js/common/core/crypto/key.js';
 import { opgp } from '../../js/common/core/crypto/pgp/openpgpjs-custom.js';
 import { Lang } from '../../js/common/lang.js';
-import { Catch, CompanyLdapKeyMismatchError } from '../../js/common/platform/catch.js';
+import { Catch } from '../../js/common/platform/catch.js';
+import { CompanyLdapKeyMismatchError } from '../../js/common/platform/error-report.js';
 import { AcctStore, AcctStoreDict } from '../../js/common/platform/store/acct-store.js';
 import { KeyStore } from '../../js/common/platform/store/key-store.js';
 import { Xss } from '../../js/common/platform/xss.js';
