+++ extension/js/common/api/flowcrypt-website.ts
@@ -2,8 +2,8 @@
 'use strict';
 
 import { Api } from './shared/api.js';
-import { Catch } from '../platform/catch.js';
 import { Browser } from '../browser/browser.js';
+import { CatchHelper } from '../platform/catch-helper.js';
 
 namespace FlowCryptWebsiteRes {
   export type FcBlogPost = { title: string; date: string; url: string };
@@ -15,7 +15,7 @@ export class FlowCryptWebsite extends Api {
   };
 
   public static retrieveBlogPosts = async (): Promise<FlowCryptWebsiteRes.FcBlogPost[]> => {
-    const xmlString = await Api.ajax({ url: 'https://flowcrypt.com/blog/feed.xml', method: 'GET', stack: Catch.stackTrace() }, 'text');
+    const xmlString = await Api.ajax({ url: 'https://flowcrypt.com/blog/feed.xml', method: 'GET', stack: CatchHelper.stackTrace() }, 'text');
     const xml = $.parseXML(xmlString);
     const posts: FlowCryptWebsiteRes.FcBlogPost[] = [];
     for (const post of Browser.arrFromDomNodeList(xml.querySelectorAll('entry'))) {
