+++ extension/js/common/assert.ts
@@ -7,7 +7,8 @@ import { Ui } from './browser/ui.js';
 import { Dict, UrlParam, UrlParams } from './core/common.js';
 import { KeyInfoWithIdentity, KeyUtil } from './core/crypto/key.js';
 import { Lang } from './lang.js';
-import { Catch, UnreportableError } from './platform/catch.js';
+import { Catch } from './platform/catch.js';
+import { UnreportableError } from './platform/error-report.js';
 import { AcctStore } from './platform/store/acct-store.js';
 import { KeyStore } from './platform/store/key-store.js';
 import { Xss } from './platform/xss.js';
